+++ Analysis/Tasks/PWGLF/spectraTPC.cxx
@@ -33,7 +33,7 @@ void customize(std::vector<o2::framework::ConfigParamSpec>& workflowOptions)
   Configurable<float> cfgCutVertex{"cfgCutVertex", 10.0f, "Accepted z-vertex range"}; \
   Configurable<float> cfgCutEta{"cfgCutEta", 0.8f, "Eta range for tracks"};           \
   Filter collisionFilter = nabs(aod::collision::posZ) < cfgCutVertex;                 \
-  Filter trackFilter = (nabs(aod::track::eta) < cfgCutEta) && (aod::track::isGlobalTrack == true);
+  Filter trackFilter = (nabs(aod::track::eta) < cfgCutEta) && (aod::track::isGlobalTrack == (uint8_t) true);
 
 #define makelogaxis(h)                                            \
   {                                                               \
